Prefer double to float for energy calculation, only add shapes if they improve score, fix a couple rasterization bugs

+++ geometrize/geometrize/shaperesult.h
@@ -18,7 +18,7 @@ namespace geometrize
  */
 struct ShapeResult
 {
-    const float score;
+    const double score;
     const geometrize::rgba color;
     const std::shared_ptr<geometrize::Shape> shape;
 };
